requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
google-cloud-texttospeech>=2.14.1
google-cloud-speech>=2.21.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0.1
requests>=2.31.0
python-dateutil>=2.8.2
uuid>=1.30
pinecone